# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gestion de Incidentes API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/incidentes/": {
      "post": {
        "tags": [
          "Incidentes"
        ],
        "summary": "Crear y clasificar un incidente",
        "description": "Crea un nuevo incidente y ejecuta la clasificación automática.\n\nInvocado por N8N tras recibir un correo electrónico en Outlook o una\ntranscripción de llamada de Twilio, o directamente desde un cliente HTTP.\n\nEl incidente se persiste en estado \"nuevo\" y es clasificado inmediatamente\npor el pipeline híbrido (determinístico + Gemini). El resultado incluye\nel sector asignado, la confianza de la clasificación y el indicador\nde revisión humana si la confianza fue insuficiente.\n\nArgs:\n    payload: Descripción pseudonimizada del incidente, prioridad y canal.\n\nReturns:\n    Representación completa del incidente creado y clasificado (HTTP 201).",
        "operationId": "create_incidente_api_v1_incidentes__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidenteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidenteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Incidentes"
        ],
        "summary": "Listar incidentes con filtros opcionales",
        "description": "Lista incidentes con filtros combinables y paginación.\n\nTodos los parámetros de filtro son opcionales y se combinan con AND lógico.\nEl resultado se ordena por fecha de creación descendente (más reciente primero).\nRetorna la proyección ligera IncidenteListItem (sin texto de descripción)\npara optimizar el tamaño de la respuesta en consultas masivas.",
        "operationId": "list_incidentes_api_v1_incidentes__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sector_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por sector responsable",
              "title": "Sector Id"
            },
            "description": "Filtrar por sector responsable"
          },
          {
            "name": "estado_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por estado del ciclo de vida",
              "title": "Estado Id"
            },
            "description": "Filtrar por estado del ciclo de vida"
          },
          {
            "name": "prioridad",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PrioridadEnum"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por nivel de prioridad",
              "title": "Prioridad"
            },
            "description": "Filtrar por nivel de prioridad"
          },
          {
            "name": "requiere_revision_humana",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por indicador de revisión pendiente",
              "title": "Requiere Revision Humana"
            },
            "description": "Filtrar por indicador de revisión pendiente"
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha de creación mínima (ISO 8601)",
              "title": "Desde"
            },
            "description": "Fecha de creación mínima (ISO 8601)"
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Fecha de creación máxima (ISO 8601)",
              "title": "Hasta"
            },
            "description": "Fecha de creación máxima (ISO 8601)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Cantidad máxima de resultados",
              "default": 50,
              "title": "Limit"
            },
            "description": "Cantidad máxima de resultados"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Desplazamiento para paginación",
              "default": 0,
              "title": "Offset"
            },
            "description": "Desplazamiento para paginación"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IncidenteListItem"
                  },
                  "title": "Response List Incidentes Api V1 Incidentes  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidentes/{incidente_id}": {
      "get": {
        "tags": [
          "Incidentes"
        ],
        "summary": "Obtener el detalle completo de un incidente",
        "description": "Retorna la representación completa de un incidente, incluyendo relaciones.\n\nA diferencia del listado, esta respuesta incluye la descripción completa\ndel incidente y los objetos de catálogo completos (sector, estado, canal).\n\nArgs:\n    incidente_id: ID del incidente a recuperar.\n\nReturns:\n    Detalle completo del incidente (HTTP 200) o error 404 si no existe.",
        "operationId": "get_incidente_api_v1_incidentes__incidente_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incidente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Incidente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidenteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Incidentes"
        ],
        "summary": "Actualizar parcialmente un incidente",
        "description": "Actualiza uno o más campos de un incidente existente (PATCH semántico).\n\nSolo se modifican los campos incluidos en el payload con valor distinto\nde None. Casos de uso típicos: cambio de estado por el operador,\nreasignación de sector tras revisión manual, o modificación de prioridad.\n\nLos incidentes en estado terminal (cerrado, es_terminal=true) son de solo\nlectura. Cualquier intento de PATCH sobre un incidente cerrado retorna\nHTTP 409 Conflict con código INCIDENTE_CERRADO.\n\nArgs:\n    incidente_id: ID del incidente a modificar.\n    payload:      Campos a actualizar (todos opcionales).\n\nReturns:\n    Representación actualizada del incidente (HTTP 200).\n    HTTP 404 si el incidente no existe.\n    HTTP 409 si el incidente está cerrado (solo lectura).",
        "operationId": "update_incidente_api_v1_incidentes__incidente_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incidente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Incidente Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidenteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidenteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/clasificaciones/incidente/{incidente_id}": {
      "get": {
        "tags": [
          "Clasificaciones"
        ],
        "summary": "Historial de clasificaciones de un incidente",
        "description": "Retorna el historial completo de clasificaciones de un incidente específico.\n\nPermite al operador ver todas las decisiones tomadas por el sistema\npara un incidente (incluyendo el registro de la etapa utilizada,\nla confianza, y la respuesta raw de Gemini si corresponde).\n\nArgs:\n    incidente_id: ID del incidente cuyo historial se desea consultar.\n\nReturns:\n    Lista de registros de auditoría ordenados del más reciente al más antiguo.",
        "operationId": "list_by_incidente_api_v1_clasificaciones_incidente__incidente_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incidente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Incidente Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClasificacionLogRead"
                  },
                  "title": "Response List By Incidente Api V1 Clasificaciones Incidente  Incidente Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanalOrigenRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "descripcion"
        ],
        "title": "CanalOrigenRead",
        "description": "Proyección de solo lectura del canal de ingreso del incidente.\n\nInforma al consumidor de la API a través de qué medio fue recibido\nel reporte: correo electrónico, formulario web o llamada telefónica."
      },
      "ClasificacionLogRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "incidente_id": {
            "type": "integer",
            "title": "Incidente Id"
          },
          "confianza": {
            "type": "number",
            "title": "Confianza"
          },
          "etapa": {
            "type": "string",
            "title": "Etapa"
          },
          "requiere_revision_humana": {
            "type": "boolean",
            "title": "Requiere Revision Humana"
          },
          "respuesta_raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Respuesta Raw"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sector_predicho": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectorRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "sector_validado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectorRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "incidente_id",
          "confianza",
          "etapa",
          "requiere_revision_humana",
          "respuesta_raw",
          "created_at",
          "sector_predicho",
          "sector_validado"
        ],
        "title": "ClasificacionLogRead",
        "description": "Representación de un registro de auditoría para la API HTTP.\n\nExpone el historial de clasificaciones de un incidente, incluyendo\nel sector predicho y el sector validado manualmente (si existe).\nEs la forma en que los operadores humanos acceden a los registros\npendientes de revisión a través de los endpoints correspondientes."
      },
...
      "EstadoRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "es_terminal": {
            "type": "boolean",
            "title": "Es Terminal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "descripcion",
          "es_terminal"
        ],
        "title": "EstadoRead",
        "description": "Proyección de solo lectura de un estado del ciclo de vida del incidente.\n\nEl campo es_terminal permite a los clientes de la API identificar si el\nincidente llegó a un estado final sin necesidad de lógica adicional."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidenteCreate": {
        "properties": {
          "descripcion": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 10,
            "title": "Descripcion",
            "description": "Texto pseudonimizado del incidente. Mínimo 10 caracteres."
          },
          "prioridad": {
            "$ref": "#/components/schemas/PrioridadEnum",
            "default": "media"
          },
          "canal_origen_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canal Origen Id"
          }
        },
        "type": "object",
        "required": [
          "descripcion"
        ],
        "title": "IncidenteCreate",
        "description": "Payload de creación de un incidente.\n\nAceptado por el endpoint POST /api/v1/incidentes. Puede provenir de:\n    - El webhook de N8N tras recibir un correo electrónico en Outlook.\n    - El webhook de N8N tras procesar la transcripción de una llamada en Twilio.\n    - Una llamada directa a la API desde un cliente externo.\n\nValidaciones de negocio:\n    - La descripción no puede estar en blanco ni ser solo espacios.\n    - La longitud mínima de 10 caracteres evita incidentes vacíos de contenido.\n    - La longitud máxima de 5000 caracteres previene abusos de almacenamiento."
      },
      "IncidenteListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "prioridad": {
            "$ref": "#/components/schemas/PrioridadEnum"
          },
          "requiere_revision_humana": {
            "type": "boolean",
            "title": "Requiere Revision Humana"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "sector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectorRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoRead"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prioridad",
          "requiere_revision_humana",
          "created_at",
          "sector",
          "estado"
        ],
        "title": "IncidenteListItem",
        "description": "Proyección ligera de un incidente para endpoints de listado.\n\nOmite el campo 'descripcion' para reducir el tamaño de las respuestas\ncuando se listan múltiples incidentes. Los consumidores que necesiten\nel texto completo deben usar el endpoint de detalle individual."
      },
      "IncidenteRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "descripcion_pseudonimizada": {
            "type": "string",
            "title": "Descripcion Pseudonimizada"
          },
          "prioridad": {
            "$ref": "#/components/schemas/PrioridadEnum"
          },
          "requiere_revision_humana": {
            "type": "boolean",
            "title": "Requiere Revision Humana"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "sector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectorRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoRead"
          },
          "canal_origen": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CanalOrigenRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "descripcion_pseudonimizada",
          "prioridad",
          "requiere_revision_humana",
          "created_at",
          "updated_at",
          "sector",
          "estado",
          "canal_origen"
        ],
        "title": "IncidenteRead",
        "description": "Representación completa de un incidente para el endpoint de detalle.\n\nIncluye los objetos relacionados completos (sector, estado, canal_origen)\nen lugar de solo sus IDs, para que el cliente no necesite realizar\nconsultas adicionales para obtener sus nombres.\n\nPolítica de privacidad (C-03, Ley 25.326):\n    Solo se expone 'descripcion_pseudonimizada' (texto con etiquetas PII).\n    'descripcion_original' (cifrada, con PII) NO forma parte de esta respuesta."
      },
      "IncidenteUpdate": {
        "properties": {
          "prioridad": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrioridadEnum"
              },
              {
                "type": "null"
              }
            ]
          },
          "estado_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estado Id"
          },
          "sector_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Id"
          },
          "requiere_revision_humana": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Revision Humana"
          }
        },
        "type": "object",
        "title": "IncidenteUpdate",
        "description": "Payload de actualización parcial de un incidente.\n\nTodos los campos son opcionales. El servicio aplica únicamente los\ncampos con valor no nulo, implementando el patrón PATCH semántico.\n\nCasos de uso típicos:\n    - Cambio de estado por el operador de mesa de ayuda.\n    - Reasignación de sector tras revisión manual.\n    - Cambio de prioridad ante escalamiento.\n    - Marcado o desmarcado de revisión humana requerida."
      },
...
      "PrioridadEnum": {
        "type": "string",
        "enum": [
          "baja",
          "media",
          "alta",
          "critica"
        ],
        "title": "PrioridadEnum",
        "description": "Enumeración de niveles de prioridad de un incidente.\n\nHereda de str para que SQLAlchemy almacene el valor textual en la base\nde datos, lo que mejora la legibilidad de consultas directas y simplifica\nla serialización en las respuestas JSON de la API.\n\nNiveles (de menor a mayor urgencia):\n    baja    → Puede ser atendido en el próximo turno disponible.\n    media   → Valor por defecto; afecta a un usuario individual.\n    alta    → Afecta a múltiples usuarios o un proceso crítico.\n    critica → Detención de operaciones; requiere atención inmediata."
      },
...
      "SectorRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "descripcion"
        ],
        "title": "SectorRead",
        "description": "Proyección de solo lectura de un sector responsable.\n\nUtilizada en las respuestas de endpoints de incidentes y clasificaciones\npara presentar el sector asignado sin exponer detalles internos del ORM."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}